+++ src/button/button.props.tsx
@@ -1,11 +1,9 @@
-import type { ButtonHTMLAttributes, ReactNode } from 'react';
+import type { ReactNode } from 'react';
 import type { CSSProp } from 'styled-components';
 import { EffectsProps } from '../effects/effects.props';
 import { Theme } from '../theming/theming';
 
-export interface ButtonProps
-  extends ButtonHTMLAttributes<HTMLButtonElement>,
-    EffectsProps {
+export interface ButtonProps extends EffectsProps {
   children: ReactNode /** @todo this should probably be more specific */;
   css?: CSSProp;
   disabled?: boolean;
@@ -16,6 +14,19 @@ export interface ButtonProps
   variant?: 'solid' | 'outline' | 'ghost' | 'link';
 }
 
-export interface ButtonTextProps {
-  children: ReactNode;
-}
+// export interface ButtonProps
+//   extends ButtonHTMLAttributes<HTMLButtonElement>,
+//     EffectsProps {
+//   children: ReactNode /** @todo this should probably be more specific */;
+//   css?: CSSProp;
+//   disabled?: boolean;
+//   shape?: 'default' | 'round' | 'square';
+//   size?: 'sm' | 'md' | 'lg';
+//   state?: 'idle' | 'pending' | 'success' | 'error';
+//   theme?: Theme | undefined;
+//   variant?: 'solid' | 'outline' | 'ghost' | 'link';
+// }
+
+// export interface ButtonTextProps {
+//   children: ReactNode;
+// }