+++ src/switch/switch.tsx
@@ -1,40 +1,17 @@
-import { useState } from 'react';
-import type { KeyboardEvent } from 'react';
-import { SwitchProps } from './switch.props';
-import { HiddenCheckboxInput, Rail, SwitchWrap, Toggle } from './switch.styles';
+import { SwitchProps as AriaSwitchProps } from 'react-aria-components';
+import { Indicator, StyledSwitch } from './switch.styles';
 
-export function Switch(props: SwitchProps) {
-  const { defaultChecked, label } = props;
-  const [isChecked, setIsChecked] = useState(defaultChecked ?? false);
+interface SwitchProps extends Omit<AriaSwitchProps, 'children'> {
+  children: React.ReactNode;
+}
 
-  const handleClick = () => {
-    setIsChecked(!isChecked);
-  };
-
-  const handleKeyDown = (e: KeyboardEvent) => {
-    if (e.key === ' ' || e.code === 'Space' || e.keyCode === 32) {
-      e.preventDefault();
-      setIsChecked(!isChecked);
-    }
-  };
-
+export function Switch({ children, ...props }: SwitchProps) {
   return (
-    <>
-      <SwitchWrap
-        onClick={handleClick}
-        onKeyDown={handleKeyDown}
-        role="checkbox"
-        aria-checked={isChecked}
-      >
-        <Rail isChecked={isChecked} tabIndex={0}>
-          <Toggle isChecked={isChecked} />
-        </Rail>
-        {label}
-      </SwitchWrap>
-
-      {/* eslint-disable-next-line react/jsx-props-no-spreading */}
-      <HiddenCheckboxInput type="checkbox" checked={isChecked} {...props} />
-    </>
+    // eslint-disable-next-line react/jsx-props-no-spreading
+    <StyledSwitch {...props}>
+      <Indicator />
+      {children}
+    </StyledSwitch>
   );
 }