+++ ui-kit/src/drawer/drawer.props.tsx
@@ -1,5 +1,0 @@
-import type { ModalProps } from '../modal/modal.props';
-
-export interface DrawerProps extends ModalProps {
-  position?: 'top' | 'right' | 'bottom' | 'left';
-}